requirements.txt
streamlit==1.37.1
openai>=1.47.0
python-dotenv>=1.0.1
audio-recorder-streamlit==0.0.10
langdetect==1.0.9
pydub==0.25.1